Finalizado processamento de sessions

+++ src/Monkify.Domain/Monkey/Entities/Bet.cs
@@ -10,7 +10,7 @@ namespace Monkify.Domain.Monkey.Entities
 {
     public class Bet : TableEntity
     {
-        public decimal BetAmount { get; set; }
+        public double BetAmount { get; set; }
         public string BetChoice { get; set; }
         public User User { get; set; }
         public Guid UserId { get; set; }
